create initial code for search bar and create list page
