+++ src/Pages/CreateList/style.module.css
@@ -1,0 +1,3 @@
+.container {
+   width: 826px;
+}
